requirements/base.txt
# Base requirements for ChatPage project
# Core Django and API dependencies

django==5.2.7
channels==4.3.1
channels_redis==4.3.0
djangorestframework==3.16.1
bleach==6.2.0
Pillow==11.0.0

# Database
psycopg2-binary==2.9.11
